rps_game/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
psycopg2-binary==2.9.9
python-consul==1.1.0
